src/requirements.txt
accelerate==0.24.1
aiohttp==3.9.0
aiosignal==1.3.1
asttokens==2.4.1
attrs==23.1.0
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.0
datasets==2.15.0
debugpy==1.8.0
decorator==5.1.1
dill==0.3.7
docopt==0.6.2
executing==2.0.1
filelock==3.13.1
frozenlist==1.4.0
fsspec==2023.10.0
huggingface-hub==0.16.4
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==8.6.0
jupyter_core==5.5.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mpmath==1.3.0
mtranslate==1.8
multidict==6.0.4
multiprocess==0.70.15
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
pipreqs==0.4.13
platformdirs==3.11.0
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
pyarrow==14.0.1
pyarrow-hotfix==0.5
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
sentencepiece==0.1.99
six==1.16.0
sklearn==0.0.post11
stack-data==0.6.3
sympy==1.12
tokenizers==0.15.0
torch==2.1.1
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
transformers==4.35.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.9
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.2
